improve cloudlab dpdk setup script

+++ scripts/setup_dpdk/cloudlab.sh
@@ -6,6 +6,11 @@ sudo modprobe uio
 sudo modprobe igb_uio
 
 # Bind CloudLab's experimental interface (10.*.*.*) to DPDK
+if [ ! -f ifname ]; then
+  echo "Error: scripts/setup-dpdk/ifname not found."
+  exit
+fi
+
 (cd $(dirname $0); rm ifname; g++ -std=c++11 -o ifname ifname.cc)
 exp_interfaces=`"$(dirname $0)"/ifname 10.`
 echo "Binding interfaces $exp_interfaces to DPDK"
